feat: restructure database and entrypoint layers - folders
