+++ src/index.ts
@@ -1,6 +1,6 @@
 import PrismaPaymentRepository from '@infra/database/postgress/prisma-payment-repository'
-import { TechChallengeAPI } from '@infra/frameworks/express/server'
-import { PrismaService } from '@infra/frameworks/prisma/prisma.service'
+import { TechChallengeAPI } from '@infra/entrypoint/express/server'
+import { PrismaService } from '@infra/database/prisma/prisma.service'
 
 const paymentDataSource = new PrismaPaymentRepository(new PrismaService())
