+++ src/infraestructure/entrypoint/express/global-error-handling.ts
